Steganography
The java files, pictures, and text files pertain to side project of mine in regards to steganography. This was a subject that was brought up in my Computer Graphics class that I thought was interesting. Steganography is the practice of concealing a message, image, video, etc. within another message, image, video, etc. These java files can take in an image and a message, and then it will use some of the least significant bits of a pixel's RGB value to place a single character that comes from the standard ASCII table (0-127). I am considering working further on this, when I get some spare time, to create a simple Android app that could perform this code, if possible.